Refactor config.json into token, config, and permLevels files
