+++ src/commands/info.js
@@ -2,7 +2,7 @@ const got = require('got');
 const path = require('path');
 const MessageEmbed = require((require.resolve('discord.js')).split(path.sep).slice(0, -1).join(path.sep) + `${path.sep}structures${path.sep}MessageEmbed.js`);
 const authorReply = require('../util/authorReply.js');
-const { prefix } = require('../components/config.js');
+const { prefix } = require('../components/config.json');
 
 module.exports = {
 	name: 'info',
